+++ src/index.ts
@@ -30,7 +30,8 @@ export {
   PopupCancelledError,
   PopupOpenError,
   MfaRequiredError,
-  MissingRefreshTokenError
+  MissingRefreshTokenError,
+  UseDpopNonceError
 } from './errors';
 
 export {
@@ -46,3 +47,5 @@ export {
   CacheKey,
   CacheKeyData
 } from './cache';
+
+export { type FetcherConfig } from './fetcher';