+++ CHANGELOG.md
@@ -1,48 +1,91 @@
 # Change Log
 
+## [v2.4.1](https://github.com/auth0/auth0-spa-js/tree/v2.4.1) (2025-09-10)
+[Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.4.0...v2.4.1)
+
+**Fixed**
+- fix: move runtime dependencies from devDependencies to dependencies [\#1411](https://github.com/auth0/auth0-spa-js/pull/1411) ([gyaneshgouraw-okta](https://github.com/gyaneshgouraw-okta))
+
+## [v2.4.0](https://github.com/auth0/auth0-spa-js/tree/v2.4.0) (2025-09-08)
+[Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.3.0...v2.4.0)
+
+**Added**
+- Add HTTP fetcher for DPoP [\#1400](https://github.com/auth0/auth0-spa-js/pull/1400) ([martinml](https://github.com/martinml))
+- Add support for DPoP [\#1388](https://github.com/auth0/auth0-spa-js/pull/1388) ([martinml](https://github.com/martinml))
+
+## [v2.3.0](https://github.com/auth0/auth0-spa-js/tree/v2.3.0) (2025-07-16)
+[Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.2.0...v2.3.0)
+
+**Fixed**
+- Fix: Token Exchange Ignoring Scope and Audience Parameters [\#1365](https://github.com/auth0/auth0-spa-js/pull/1365) ([tusharpandey13](https://github.com/tusharpandey13))
+- bugfix: Correctly extract origin from domainUrl [\#1372](https://github.com/auth0/auth0-spa-js/pull/1372) ([tusharpandey13](https://github.com/tusharpandey13))
+
+## [v2.2.0](https://github.com/auth0/auth0-spa-js/tree/v2.2.0) (2025-05-28)
+
+[Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.1.3...v2.2.0)
+
+**Added**
+
+- Custom Token Exchange [\#1344](https://github.com/auth0/auth0-spa-js/pull/1344) ([tusharpandey13](https://github.com/tusharpandey13))
+
 ## [v2.1.3](https://github.com/auth0/auth0-spa-js/tree/v2.1.3) (2023-12-11)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.1.2...v2.1.3)
 
 **Changed**
+
 - feat: support for hosted token worker [\#1208](https://github.com/auth0/auth0-spa-js/pull/1208) ([DJMcK](https://github.com/DJMcK))
 
 ## [v2.1.2](https://github.com/auth0/auth0-spa-js/tree/v2.1.2) (2023-08-21)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.1.1...v2.1.2)
 
 **Fixed**
+
 - Ensure organization cookie is set when no organization was set [\#1123](https://github.com/auth0/auth0-spa-js/pull/1123) ([frederikprijck](https://github.com/frederikprijck))
 
 ## [v2.1.1](https://github.com/auth0/auth0-spa-js/tree/v2.1.1) (2023-07-18)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.1.0...v2.1.1)
 
 **Changed**
+
 - Do not lowercase org_name claim [\#1117](https://github.com/auth0/auth0-spa-js/pull/1117) ([frederikprijck](https://github.com/frederikprijck))
 
 ## [v2.1.0](https://github.com/auth0/auth0-spa-js/tree/v2.1.0) (2023-07-13)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.0.8...v2.1.0)
 
 **Added**
+
 - Support Organization Name [\#1113](https://github.com/auth0/auth0-spa-js/pull/1113) ([frederikprijck](https://github.com/frederikprijck))
 
 **Fixed**
+
 - Ensure AMR claim is set to an array of strings [\#1112](https://github.com/auth0/auth0-spa-js/pull/1112) ([frederikprijck](https://github.com/frederikprijck))
 
 ## [v2.0.8](https://github.com/auth0/auth0-spa-js/tree/v2.0.8) (2023-06-14)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.0.7...v2.0.8)
 
 **Changed**
+
 - Lazily retrieve transaction from transaction storage [\#1108](https://github.com/auth0/auth0-spa-js/pull/1108) ([frederikprijck](https://github.com/frederikprijck))
 
 ## [v2.0.7](https://github.com/auth0/auth0-spa-js/tree/v2.0.7) (2023-06-02)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.0.6...v2.0.7)
 
 **Changed**
+
 - Make TransactionManager use CookieDomain [\#1105](https://github.com/auth0/auth0-spa-js/pull/1105) ([ZdravkoDonev-gtmhub](https://github.com/ZdravkoDonev-gtmhub))
 
 ## [v2.0.6](https://github.com/auth0/auth0-spa-js/tree/v2.0.6) (2023-05-30)
+
 [Full Changelog](https://github.com/auth0/auth0-spa-js/compare/v2.0.5...v2.0.6)
 
 **Fixed**
+
 - Fix missing invalid state errors with Generic Error [\#1102](https://github.com/auth0/auth0-spa-js/pull/1102) ([frederikprijck](https://github.com/frederikprijck))
 
 ## [v2.0.5](https://github.com/auth0/auth0-spa-js/tree/v2.0.5) (2023-05-22)